+++ src/model_state_sort.h
@@ -5,7 +5,7 @@
 
 struct model_state;
 
-void model_state_name_sort(struct model_state const** mstates, uint32_t nstates);
-int  model_state_topological_sort(struct model_state const** mstates, uint32_t nstates);
+void model_state_name_sort(struct model_state const** mstates, uint16_t nstates);
+int  model_state_topological_sort(struct model_state const** mstates, uint16_t nstates);
 
 #endif
