+++ src/model_trans_table.h
@@ -1,6 +1,8 @@
 #ifndef MODEL_TRANS_TABLE_H
 #define MODEL_TRANS_TABLE_H
 
+#include <stdint.h>
+
 struct imm_state;
 struct model_trans;
 struct model_trans_table;
@@ -15,7 +17,7 @@ unsigned long              model_trans_table_end(struct model_trans_table const*
 int                        model_trans_table_exist(struct model_trans_table const* table, unsigned long iter);
 unsigned long              model_trans_table_find(struct model_trans_table const* table, struct imm_state const* state);
 struct model_trans*        model_trans_table_get(struct model_trans_table const* table, unsigned long iter);
-unsigned                   model_trans_table_size(struct model_trans_table const* table);
+uint16_t                   model_trans_table_size(struct model_trans_table const* table);
 
 /** iterate over buckets */
 #define model_trans_table_for_each(iter, tbl)                                                                          \
